#7680: move registry types to registryTypes
